+++ examples/react-native-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @knocklabs/react-native-example
 
+## 1.0.4
+
+### Patch Changes
+
+- @knocklabs/react-native@0.1.4
+
 ## 1.0.3
 
 ### Patch Changes