+++ packages/react-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.6
+
+### Patch Changes
+
+- c9faba5: fix esm build issues with mjs files
+- Updated dependencies [c9faba5]
+  - @knocklabs/client@0.8.21
+
 ## 0.1.5
 
 ### Patch Changes